refactor header to be in a header folder, so that we can make the main navbar a server component
